Code Changes: Package Refactor and Description Update
