--- src/aws_resource_manager/ec2/__init__.py
@@ -1,1 +1,0 @@
-"""python package containing the ec2 handler module."""
