--- src/aws_resource_manager/redshift/_init__.py
@@ -1,1 +1,0 @@
-"""python package containing the redshift handler module."""
