+++ src/aws_resource_manager/s3.py
@@ -1,3 +1,4 @@
+"""Module for managing AWS S3 resources."""
 import io
 import os
 from io import StringIO
@@ -5,8 +6,7 @@ from io import StringIO
 import boto3
 from botocore.exceptions import ClientError
 from dotenv import load_dotenv
-
-from src.aws_resource_manager.s3.logger import configure_logger
+from logger import configure_logger
 
 logger = configure_logger()
