--- src/aws_resource_manager/s3/__init__.py
@@ -1,1 +1,0 @@
-"""python package containing the s3 handler module."""
